Dropped Oath2 namespaced OMSDK - switched to Verizonmedia
