--- OMSDK_Oath2.framework/Headers
@@ -1,1 +1,0 @@
-Versions/Current/Headers
