--- OMSDK_Oath2.framework/Modules/module.modulemap
@@ -1,4 +1,0 @@
-module OMSDK_Oath2 {
-    header "Headers/OMIDImports.h"
-    export *
-}
