--- OMSDK_Oath2.framework/OMSDK_Oath2
@@ -1,1 +1,0 @@
-Versions/Current/OMSDK_Oath2
